requirements.txt
# Web Framework
flask==3.0.2
python-dotenv==1.0.1

# Database
mysql-connector-python==8.3.0

# AI/ML Components
langchain>=0.1.12
langchain-openai>=0.0.8
langchain-community>=0.0.28
langchain-experimental>=0.0.49
langchain-pinecone>=0.0.1
sentence-transformers>=2.5.1
pinecone-client[grpc]>=3.0.2

# Document Processing
pypdf>=4.1.0

# Development Tools
pytest>=8.0.2
black>=24.2.0
flake8>=7.0.0

# Local Package
-e .